refactor: improve structure and use native httpStatusCodes from node
