+++ src/openapi/schemas/params-schema.ts
@@ -11,3 +11,5 @@ export const idParamsSchema = z.object({
     example: 1,
   }),
 });
+
+
